Ahora sí, están los sql2o en detalleorden y producto.
También complete el crud respectivo de cada uno.

+++ src/main/java/tbd/lab1/entities/CategoriaEntity.java
@@ -4,7 +4,7 @@ import lombok.Data;
 @Data
 public class CategoriaEntity {
 
-    private Long id_categoria;
+    private Integer id_categoria;
 
     private String nombre;
 }
